Changes to login, register and tabs
